+++ GPTifier/include/commands.hpp
