+++ GPTifier/src/commands.cpp
@@ -1,4 +1,4 @@
-#include "run_completion.hpp"
+#include "commands.hpp"
 
 #include "params.hpp"
 #include "prompts.hpp"
