+++ GPTifier/src/main.cpp
@@ -1,5 +1,5 @@
+#include "commands.hpp"
 #include "params.hpp"
-#include "run_completion.hpp"
 
 #include <iostream>
 #include <stdexcept>
